reformat basket page structure
