+++ client/src/features/basket/BasketPage.tsx
@@ -1,128 +1,24 @@
 import {
-  Box,
   Button,
   Grid,
-  Paper,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
   Typography,
 } from "@mui/material";
-import { currencyFormat } from "../../app/helpers/utils";
-import { Add, Delete, Remove } from "@mui/icons-material";
-import { LoadingButton } from "@mui/lab";
 import BasketSummary from "./BasketSummary";
 import { Link } from "react-router-dom";
 import { useAppSelector } from "../../app/hooks/useAppSelector";
-import { basketSelector, statusSelector } from "../../state/basket/selectors";
-import { useAppDispatch } from "../../app/hooks/useAppDispatch";
-import {
-  addBasketItemAsync,
-  removeBasketItemAsync,
-} from "../../state/basket/actions";
+import { basketSelector } from "../../state/basket/selectors";
+import BasketTable from "./BasketTable";
 
 export default function BasketPage() {
   const basket = useAppSelector(basketSelector);
-  const {status} = useAppSelector(statusSelector);
-  const dispatch = useAppDispatch();
 
   if (!basket)
     return <Typography variant="h3">Your basket is empty</Typography>;
 
   return (
     <>
-      <TableContainer component={Paper}>
-        <Table sx={{ minWidth: 650 }}>
-          <TableHead>
-            <TableRow>
-              <TableCell>Product</TableCell>
-              <TableCell align="right">Price</TableCell>
-              <TableCell align="center">Quantity</TableCell>
-              <TableCell align="right">Subtotal</TableCell>
-              <TableCell align="right"></TableCell>
-            </TableRow>
-          </TableHead>
-
-          <TableBody>
-            {basket.items.map((item) => (
-              <TableRow
-                key={item.productId}
-                sx={{ "&:last-child td, &:last-child th": { border: 0 } }}
-              >
-                <TableCell component="th" scope="row">
-                  <Box display="flex" alignItems="center">
-                    <img
-                      src={item.pictureUrl}
-                      alt={item.name}
-                      style={{ height: 50, marginRight: 20 }}
-                    />
-                    <span>{item.name}</span>
-                  </Box>
-                </TableCell>
-                <TableCell align="right">
-                  {currencyFormat(item.price)}
-                </TableCell>
-                <TableCell align="center">
-                  <LoadingButton
-                    loading={
-                      status === "pendingRemoveItem" + item.productId + "rem"
-                    }
-                    onClick={() =>
-                      dispatch(
-                        removeBasketItemAsync({
-                          productId: item.productId,
-                          quantity: 1,
-                          name: "rem",
-                        })
-                      )
-                    }
-                    color="error"
-                  >
-                    <Remove />
-                  </LoadingButton>
-                  {item.quantity}
-                  <LoadingButton
-                    loading={status === "pendingAddItem" + item.productId}
-                    onClick={() =>
-                      dispatch(
-                        addBasketItemAsync({ productId: item.productId })
-                      )
-                    }
-                    color="secondary"
-                  >
-                    <Add />
-                  </LoadingButton>
-                </TableCell>
-                <TableCell align="right">
-                  ${((item.price / 100) * item.quantity).toFixed(2)}
-                </TableCell>
-                <TableCell align="right">
-                  <LoadingButton
-                    loading={
-                      status === "pendingRemoveItem" + item.productId + "del"
-                    }
-                    onClick={() =>
-                      dispatch(
-                        removeBasketItemAsync({
-                          productId: item.productId,
-                          quantity: item.quantity,
-                          name: "del",
-                        })
-                      )
-                    }
-                    color="error"
-                  >
-                    <Delete />
-                  </LoadingButton>
-                </TableCell>
-              </TableRow>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
+      <BasketTable items={basket.items} />
+      
       <Grid container>
         <Grid item xs={6} />
         <Grid item xs={6}>
