+++ client/src/features/basket/BasketTable.tsx
@@ -1,0 +1,124 @@
+import { Remove, Add, Delete } from "@mui/icons-material";
+import { LoadingButton } from "@mui/lab";
+import {
+  TableContainer,
+  Paper,
+  Table,
+  TableHead,
+  TableRow,
+  TableCell,
+  TableBody,
+  Box,
+} from "@mui/material";
+import { currencyFormat } from "../../app/helpers/utils";
+import {
+  removeBasketItemAsync,
+  addBasketItemAsync,
+} from "../../state/basket/actions";
+import { BasketItem } from "../../app/models/basket";
+import { useAppSelector } from "../../app/hooks/useAppSelector";
+import { useAppDispatch } from "../../app/hooks/useAppDispatch";
+
+interface Props {
+  items: BasketItem[];
+  isBasket?: boolean;
+}
+export default function BasketTable({ items, isBasket = true }: Props) {
+  const dispatch = useAppDispatch();
+  const { status } = useAppSelector((state) => state.basket);
+  return (
+    <TableContainer component={Paper}>
+      <Table sx={{ minWidth: 650 }}>
+        <TableHead>
+          <TableRow>
+            <TableCell>Product</TableCell>
+            <TableCell align="right">Price</TableCell>
+            <TableCell align="center">Quantity</TableCell>
+            <TableCell align="right">Subtotal</TableCell>
+            {isBasket && <TableCell align="right"></TableCell>}
+          </TableRow>
+        </TableHead>
+
+        <TableBody>
+          {items.map((item) => (
+            <TableRow
+              key={item.productId}
+              sx={{ "&:last-child td, &:last-child th": { border: 0 } }}
+            >
+              <TableCell component="th" scope="row">
+                <Box display="flex" alignItems="center">
+                  <img
+                    src={item.pictureUrl}
+                    alt={item.name}
+                    style={{ height: 50, marginRight: 20 }}
+                  />
+                  <span>{item.name}</span>
+                </Box>
+              </TableCell>
+              <TableCell align="right">{currencyFormat(item.price)}</TableCell>
+              <TableCell align="center">
+                {isBasket && (
+                  <LoadingButton
+                    loading={
+                      status === "pendingRemoveItem" + item.productId + "rem"
+                    }
+                    onClick={() =>
+                      dispatch(
+                        removeBasketItemAsync({
+                          productId: item.productId,
+                          quantity: 1,
+                          name: "rem",
+                        })
+                      )
+                    }
+                    color="error"
+                  >
+                    <Remove />
+                  </LoadingButton>
+                )}
+                {item.quantity}
+                {isBasket && (
+                  <LoadingButton
+                    loading={status === "pendingAddItem" + item.productId}
+                    onClick={() =>
+                      dispatch(
+                        addBasketItemAsync({ productId: item.productId })
+                      )
+                    }
+                    color="secondary"
+                  >
+                    <Add />
+                  </LoadingButton>
+                )}
+              </TableCell>
+              <TableCell align="right">
+                ${((item.price / 100) * item.quantity).toFixed(2)}
+              </TableCell>
+              {isBasket && (
+                <TableCell align="right">
+                  <LoadingButton
+                    loading={
+                      status === "pendingRemoveItem" + item.productId + "del"
+                    }
+                    onClick={() =>
+                      dispatch(
+                        removeBasketItemAsync({
+                          productId: item.productId,
+                          quantity: item.quantity,
+                          name: "del",
+                        })
+                      )
+                    }
+                    color="error"
+                  >
+                    <Delete />
+                  </LoadingButton>
+                </TableCell>
+              )}
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
+    </TableContainer>
+  );
+}
